pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "support-orchestrator-demo"
version = "0.1.0"
description = "Thenvoi support orchestrator demo - customer support multi-agent workflow"
requires-python = ">=3.11"
authors = [{ name = "Thenvoi" }]
license = { text = "MIT" }

dependencies = [
    # Thenvoi SDK with Claude SDK extras
    "thenvoi-sdk[claude_sdk]",

    # HTTP client
    "httpx>=0.24.0",

    # Configuration
    "python-dotenv>=1.1.1",
    "pyyaml>=6.0",

    # Excel support (for ExcelAgent and demo data generation)
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",

    # Logging (colorama for colored test output)
    "colorama>=0.4.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
]

[tool.setuptools.packages.find]
where = ["src"]
include = [
    "orchestrator",
    "orchestrator.*",
    "agents",
    "agents.*",
    "thenvoi_integration",
    "thenvoi_integration.*",
]
namespaces = false

[tool.setuptools]
package-dir = {"" = "src"}

[tool.pytest.ini_options]
testpaths = ["src/tests"]
asyncio_mode = "auto"
addopts = "-v"

